feature: env validations

+++ src/user/user.controller.ts
@@ -1,4 +1,5 @@
 import {
+  BadRequestException,
   Body,
   Controller,
   Delete,
@@ -8,15 +9,16 @@ import {
   Put,
 } from "@nestjs/common";
 import { UserService } from "@/user/user.service";
-import { User } from "@/user/user.schema";
-import { userDto } from "@/user/user.dto";
+import { User } from "@/schemas/user.schema";
+import { createUserDto, updateUserDto } from "@/dtos/user.dto";
+import { isMongoIdValid } from "@/dtos/misc.dto";
 
 @Controller("user")
 export class UserController {
   constructor(private userService: UserService) {}
 
   @Post("new")
-  async createNewUser(@Body() userData: userDto): Promise<User> {
+  async createNewUser(@Body() userData: createUserDto): Promise<User> {
     return this.userService.createNewUser(userData);
   }
 
@@ -27,19 +29,31 @@ export class UserController {
 
   @Get(":userId")
   async findUserById(@Param("userId") userId: string): Promise<User> {
+    const isUserIdValid: boolean = isMongoIdValid(userId);
+    if (!isUserIdValid) {
+      throw new BadRequestException("Invalid user id!");
+    }
     return this.userService.findUserById(userId);
   }
 
   @Put(":userId")
   async updateUserById(
     @Param("userId") userId: string,
-    @Body() userData: userDto
+    @Body() userData: updateUserDto
   ): Promise<User> {
+    const isUserIdValid: boolean = isMongoIdValid(userId);
+    if (!isUserIdValid) {
+      throw new BadRequestException("Invalid user id!");
+    }
     return this.userService.updateUserById(userId, userData);
   }
 
   @Delete(":userId")
   async deleteUserById(@Param("userId") userId: string): Promise<User> {
+    const isUserIdValid: boolean = isMongoIdValid(userId);
+    if (!isUserIdValid) {
+      throw new BadRequestException("Invalid user id!");
+    }
     return this.userService.deleteUserById(userId);
   }
 }
